+++ src/plone/restapi/services/workflow/info.py
@@ -23,6 +23,11 @@ class WorkflowInfo(Service):
                 'title': action['title'],
             })
 
+        for item, action in enumerate(history):
+            history[item]['title'] = wftool.getTitleForStateOnType(
+                action['review_state'],
+                self.context.portal_type)
+
         return {
             'history': json_compatible(history),
             'transitions': transitions,